Added @draw_procs to Window and hooked them into the drawables code. This lets the bot programmer hack into the GUI and add extra visualizations. I used it to see my radar, heading, turret heading sensor, and turret heading command, and calculated target position as seen in http://www.zenspider.com/~ryan/rtanque_visualization.mov
Yes... this is a HACK. The user needs to use ObjectSpace or some other
hack to get to the Window to begin with. But this is only meant for
development so I didn't think it was that big of a deal. By caching
the Window in the bot it doesn't seem to affect performance (depending
on how much work the procs do).