Fixed collision detection when a bullet passes through a bot.
The algorithm uses the projection of the vector AB (from bullets
previous location to center of bot) onto AC (from bullets previous
location to new location). If the distance from the projection to the
center of the bot is less than the radius of the bot somewhere along the
way the bullet passed through the bot, resulting in a hit. If the
magnitude of this projection is greater than the magnitude of the
bullets projection then the bullet has not passed through the target
because the bullet has not yet traveled far enough.

Test cases have been added that describe the above.