Changes effecting maximum gun energy
Changed the bot.gun_energy_max in the configuration.rb file to be based
on the energy required to fire one maximum power shell.
Changed the initial gun energy to 0 so that there is a build up time at
the beginning of the game.
Changed the criteria for firing so that the gun energy will never go
below zero.